requirements.txt
Django>=3.2,<4.0
gdal==3.2.2.1
pyproj
psycopg2
djangorestframework>=3.14.0
djangorestframework-gis==0.16.0
django-leaflet>=0.30.1
requests>=2.28.1
django-crispy-forms==2.0.0
geopy